moved all types and interfaces into @/lib/types/index.ts file to improve code readability
