+++ src/lib/models/Note.ts
@@ -1,11 +1,5 @@
-import { Schema, ObjectId, Document } from "mongoose"
-
-export interface INote extends Document {
-	_id: ObjectId
-	content: string
-	createdAt: Date
-	createdBy: string // reference to the user (employee) who created the note - uses uuid v4
-}
+import { Schema } from "mongoose"
+import { INote } from "@/lib/types"
 
 // Note Schema - subdocument of Booking Schema
 export const noteSchema = new Schema<INote>(
